cr: #DRS-177 Removed using the database id.

+++ drs/src/core/com/db4o/drs/versant/VodCobraFacade.java
@@ -53,7 +53,7 @@ public interface VodCobraFacade {
 	
 	long queryForMySignatureLoid();
 	
-	byte[] signatureBytes(int databaseId);
+	byte[] signatureBytes(long databaseId);
 
 	
 }
